Extract functions defining dicts that specify UI names for integrator, noise and equation choices. Add an Enum to map a subform type to the proper function.
