+++ framework_tvb/tvb/interfaces/web/controllers/spatial/surface_stimulus_controller.py
@@ -39,7 +39,9 @@ import cherrypy
 import numpy
 from tvb.adapters.creators.stimulus_creator import *
 from tvb.adapters.datatypes.h5.patterns_h5 import StimuliSurfaceH5
-from tvb.adapters.simulator.equation_forms import *
+from tvb.adapters.simulator.equation_forms import get_form_for_equation
+from tvb.adapters.simulator.subforms_mapping import get_ui_name_to_equation_dict, GAUSSIAN_EQUATION, \
+    DOUBLE_GAUSSIAN_EQUATION, SIGMOID_EQUATION
 from tvb.core.entities.load import try_get_last_datatype
 from tvb.core.neocom import h5
 from tvb.core.neotraits.forms import Form, SimpleFloatField
